backend/requirements.txt
# Core Framework
fastapi==0.109.1  # Updated: Fixed ReDoS vulnerability (was 0.109.0)
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22  # Updated: Fixed multiple vulnerabilities (was 0.0.6)
authlib==1.6.6  # Updated: Fixed algorithm confusion and DoS vulnerabilities (was 1.3.0)
httpx==0.26.0

# Payment Processing
stripe==7.11.0

# Email & Mail Processing
aiosmtplib==3.0.1
aiohttp==3.13.3  # Updated: Fixed zip bomb, DoS, and directory traversal vulnerabilities (was 3.9.1)
aioimaplib==1.0.1
email-validator==2.1.0.post1

# Gmail API (for direct email injection)
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Job Queue & Cache
celery==5.3.6
redis==5.0.1

# Security & Encryption
cryptography==46.0.5  # Updated: Fixed NULL pointer dereference (was 42.0.0)

# Notifications
apprise==1.7.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.6.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
tenacity==8.2.3

# Legacy support (for migration)
poplib3==0.0.4
